requirements.txt
opencv-contrib-python==4.10.0.84
numpy==2.0.1
fastapi==0.121.3
uvicorn==0.38.0
pydantic==2.12.4
requests==2.32.3
pillow==11.0.0
python-multipart==0.0.20
transformers==4.57.3
tokenizers==0.22.1
huggingface-hub==0.36.0
regex==2025.11.3
safetensors==0.7.0 
torch==2.9.1
multipart
